Time: 2 ms (5.47%) | Memory: 8.8 MB (7.40%) - LeetSync

+++ 2010-check-if-word-equals-summation-of-two-words/check-if-word-equals-summation-of-two-words.cpp
@@ -1,13 +1,13 @@
 class Solution {
 public:
-    long long summation(const string &s)
+    int summation(const string &s)
     {
         string result = "";
         for(char c : s)
         {
             result += to_string(c-'a');
         }
-        return stoll(result);
+        return stoi(result);
     }
     bool isSumEqual(string firstWord, string secondWord, string targetWord) {
         return (summation(firstWord) + summation(secondWord) == summation(targetWord));
